// ui/src/lib/api/schema.ts
export interface components {
    schemas: {
// ...
        DriftEntry: {
            /** Symbol */
            symbol: string;
            /** Broker Qty */
            broker_qty: number;
            /** Cockpit Qty */
            cockpit_qty: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}